requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
pydantic==2.9.2
llama-cpp-python==0.3.2
langdetect==1.0.9
